Added a Scripts version of the recipe.
This uses distlib to create scripts.
Currently very hardcoded to simply make a copy of ./bin/instance to ./instance,
with a shebang line that uses the python interpreter from the plone part,
which is the one that uses pip.
And now we can use './instance fg' and it starts up a proper Plone site.

Well, that works as long as you don't have a 'dash-S' able main python,
because the parts/plone/site.py that this would generate,
conflicts with what pip and distlib try to do.

So it's a bit iffy, but under the proper circumstances and some hardcoding,
I now have a Plone instance script that uses pip-installed wheels from a virtualenv.